move config dependencies to the common module
