+++ src/pixelator/common/config/__init__.py
@@ -1,0 +1,34 @@
+"""Copyright © 2025 Pixelgen Technologies AB."""
+
+from pixelator.common.config.assay import (
+    Assay,
+    AssayModel,
+    Region,
+    RegionModel,
+    RegionType,
+    SequenceType,
+    get_position_in_parent,
+)
+from pixelator.common.config.config_class import Config, load_assays_package
+from pixelator.common.config.panel import (
+    AntibodyPanel,
+    AntibodyPanelMetadata,
+    load_antibody_panel,
+)
+from pixelator.common.config.utils import load_yaml_file
+
+__all__ = [
+    "AssayModel",
+    "RegionModel",
+    "RegionType",
+    "SequenceType",
+    "Assay",
+    "Region",
+    "get_position_in_parent",
+    "Config",
+    "load_assays_package",
+    "load_yaml_file",
+    "AntibodyPanel",
+    "AntibodyPanelMetadata",
+    "load_antibody_panel",
+]
