+++ src/pixelator/mpx/config/__init__.py
@@ -1,36 +1,13 @@
-"""Copyright © 2023 Pixelgen Technologies AB."""
+"""MPX configuration.
 
-from pixelator.mpx.config.assay import (
-    Assay,
-    AssayModel,
-    Region,
-    RegionModel,
-    RegionType,
-    SequenceType,
-    get_position_in_parent,
-)
-from pixelator.mpx.config.config_class import Config, load_assays_package
+Copyright © 2025 Pixelgen Technologies AB.
+"""
+
+from pixelator.common.config import get_position_in_parent, load_antibody_panel
 from pixelator.mpx.config.config_instance import config
-from pixelator.mpx.config.panel import (
-    AntibodyPanel,
-    AntibodyPanelMetadata,
-    load_antibody_panel,
-)
-from pixelator.mpx.config.utils import load_yaml_file
 
 __all__ = [
-    "AssayModel",
-    "RegionModel",
-    "RegionType",
-    "SequenceType",
-    "Assay",
-    "Region",
-    "get_position_in_parent",
-    "Config",
     "config",
-    "load_assays_package",
-    "load_yaml_file",
-    "AntibodyPanel",
-    "AntibodyPanelMetadata",
+    "get_position_in_parent",
     "load_antibody_panel",
 ]
