+++ src/pixelator/mpx/report/common/workdir.py
@@ -11,8 +11,8 @@ from collections import defaultdict
 from pathlib import Path, PurePath
 from typing import Dict, List, Optional, Tuple
 
+from pixelator.common.types import PathType
 from pixelator.common.utils import flatten, get_sample_name
-from pixelator.mpx.types import PathType
 
 logger = logging.getLogger(__name__)
