+++ src/pixelator/pna/config/assay.py
@@ -17,8 +17,8 @@ except ImportError:
 import pydantic
 from pydantic import BaseModel
 
-from pixelator.mpx.config.utils import load_yaml_file
-from pixelator.mpx.types import PathType
+from pixelator.common.config.utils import load_yaml_file
+from pixelator.common.types import PathType
 
 
 class PNARegionType(str, enum.Enum):
